Decompile func0f0bb69c

+++ src/include/game/game_0b63b0.h
@@ -78,7 +78,7 @@ u32 func0f0c1f08(void);
 void func0f0c1ff4(void);
 void propPlayerGetBbox(struct prop *prop, f32 *width, f32 *ymax, f32 *ymin);
 f32 currentPlayerGetHealth(void);
-u32 func0f0c21a4(void);
+f32 func0f0c21a4(void);
 void func0f0c2210(f32 shield);
 s32 getMissionTime(void);
 u32 func0f0c228c(struct prop *prop);
